// src/shared/interfaces/property/property.interfaces.ts
export type PropertyPhotoDto = {
  publicId: string;
};

export type MinimumLeaseFeeDto = {
  type: MINIMUM_LEASE_TERM_TYPE;
  value: number;
};

export type MinimumLeaseFeeDef = {
  type: MINIMUM_LEASE_TERM_TYPE;
  value: number;
};
// ...
export type PropertyAddressDef = {
  residenceComplex?: string | null;
  street: string;
  streetNumber?: string | null;
  city: string;
  postalCode: string;
  country: string;
  county: string;
  latitude: number;
  longitude: number;
  floor: string;
};

export type PropertyStatusType = "ACTIVE" | "PENDING" | "DELETED" | "EXPIRED";
// ...
export type PropertyTypeDef = {
  id: string;
  relationType: RELATION_TYPE;
  price: CurrencyType;
  description?: string | null;
  expensesMonthly: CurrencyType;
  contactPreference: CONTACT_TYPE;
  details: {
    propertyType: PROPERTY_TYPE;
    conditions: PROPERTY_CONDITIONS_TYPE;
    address: PropertyAddressDef;
    isLastFloor: boolean;
    surface: number;
    roomNumber: number;
    bathroomNumber: number;
    hasElevator: boolean;
    furniture: PROPERTY_FURNITURE_TYPE;
    residentialComplex: string | null;
    commodities: {
      hasAirConditioning: boolean;
      hasClosetInTheWall: boolean;
      hasBalcony: boolean;
      hasCellar: boolean;
      hasParking: boolean;
      hasTerrace: boolean;
    };
  };
  contractDetails: {
    maximumNumberOfTenants: number;
    petFriendly: boolean;
    agencyFee: CurrencyType;
    minimumLeaseTerm: MinimumLeaseFeeDef;
    rentInAdvance: string;
  };
  user: {
    id: string;
    firstName?: string;
    lastName?: string;
    picture: string | null;
    phoneNumbers: {
      countryCode: string;
      number: string;
      isPrimary: boolean;
    }[];
  };
  isFavorite: boolean;
  views: number;
  status: PropertyStatusType;
  photos: PropertyPhotoDto[];
  createdAt: string;
  updatedAt: string;
};